+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 
 - Support decaying rules. Decaying rules are regular sampling rules, but they are only applicable in a specific time range. ([#1544](https://github.com/getsentry/relay/pull/1544))
 - Disallow `-` in measurement and breakdown names. These items are converted to metrics, which do not allow `-` in their name. ([#1571](https://github.com/getsentry/relay/pull/1571))
-- Add `app.app_in_foreground` and `thread.ui` flag to protocol. ([#1578](https://github.com/getsentry/relay/pull/1578))
+- Add `app.in_foreground` and `thread.ui` flag to protocol. ([#1578](https://github.com/getsentry/relay/pull/1578))
 
 **Bug Fixes**: